--- a/QsNet/QsNet.csproj
+++ b/QsNet/QsNet.csproj
@@ -49,12 +49,7 @@
         <None Include="..\LICENSE" Pack="true" PackagePath="\"/>
         <None Include="..\logo.png" Pack="true" PackagePath="\"/>
         <InternalsVisibleTo Include="QsNet.Tests"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
-            <_Parameter1>QsNet.Benchmarks</_Parameter1>
-        </AssemblyAttribute>
+        <InternalsVisibleTo Include="QsNet.Benchmarks"/>
     </ItemGroup>
 
     <ItemGroup>