--- a/benchmarks/QsNet.Benchmarks/DecodeBenchmarks.cs
+++ b/benchmarks/QsNet.Benchmarks/DecodeBenchmarks.cs
@@ -3,7 +3,67 @@ using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Jobs;
 using QsNet;
 using QsNet.Models;
+using BenchmarkDotNet.Columns;
+using BenchmarkDotNet.Configs;
+using BenchmarkDotNet.Reports;
+using BenchmarkDotNet.Running;
 
+// Custom column that shows mean time per key=value pair (in ns)
+internal sealed class NsPerPairColumn : IColumn
+{
+    public string Id => nameof(NsPerPairColumn);
+    public string ColumnName => "ns/pair";
+    public string Legend => "Mean nanoseconds per parsed pair (Mean/Count)";
+    public ColumnCategory Category => ColumnCategory.Statistics;
+    public int PriorityInCategory => 100;
+    public bool AlwaysShow => true;
+    public bool IsNumeric => true;
+    public UnitType UnitType => UnitType.Time;
+
+    public string GetValue(Summary summary, BenchmarkCase benchmarkCase)
+        => GetValue(summary, benchmarkCase, summary.Style);
+
+    public string GetValue(Summary summary, BenchmarkCase benchmarkCase, SummaryStyle style)
+    {
+        var report = summary[benchmarkCase];
+        var stats = report?.ResultStatistics;
+        if (stats is null)
+            return "-";
+
+        // Benchmark parameter Count (number of key=value pairs)
+        int count = 1;
+        var parameters = benchmarkCase.Parameters.Items;
+        for (int i = 0; i < parameters.Count; i++)
+        {
+            var p = parameters[i];
+            if (p.Name == nameof(DecodeBenchmarks.Count))
+            {
+                count = p.Value is int iv ? iv : Convert.ToInt32(p.Value);
+                break;
+            }
+        }
+
+        if (count <= 0) count = 1;
+
+        // Mean is stored in nanoseconds
+        double nsPerPair = stats.Mean / count;
+        return nsPerPair.ToString("0.##");
+    }
+
+    public bool IsAvailable(Summary summary) => true;
+    public bool IsDefault(Summary summary, BenchmarkCase benchmarkCase) => false;
+}
+
+// Attach our custom column via config
+internal sealed class BenchConfig : ManualConfig
+{
+    public BenchConfig()
+    {
+        AddColumn(new NsPerPairColumn());
+    }
+}
+
+[Config(typeof(BenchConfig))]
 // Run on the build TFM by default; add others via --runtimes when installed
 #if NET8_0
 [SimpleJob(RuntimeMoniker.Net80, baseline: true)]