This pull request introduces a comprehensive benchmarking suite for the `QsNet` library and improves the performance and correctness of query string parsing, especially for comma-separated list values. The most significant changes include the addition of a dedicated benchmarking project with custom metrics, enhancements to the internal decoder logic for more efficient and accurate parsing, and minor project configuration updates.

### Benchmarking Suite Addition

* Added a new `QsNet.Benchmarks` project using BenchmarkDotNet, including a custom column (`NsPerPairColumn`) that reports nanoseconds per parsed key-value pair, and a `DecodeBenchmarks` class that tests various parsing scenarios and options. [[1]](diffhunk://#diff-b1c7fbf819ca2336b4017b82e6077f36935f44eaa2dd541c7dfa12b5708afeb3R1-R169) [[2]](diffhunk://#diff-3711ad0a4d97bf6f4ffbc18e4ef7564a7fb187990eb00787acb7170dd63a4b5eR1-R18)
* Implemented a default benchmark runner in `Program.cs` to execute all benchmarks if no filter is specified, and enabled memory diagnostics for all runs.

### Query String Parsing Improvements

* Optimized comma-separated list parsing in `Decoder.cs` to avoid unnecessary string splits and enforce list limits during parsing, improving both performance and correctness.
* Improved percent-encoding replacement logic by only scanning for replacements when a percent character is present, reducing unnecessary string operations.

### Internal Decoder Logic Enhancements

* Refactored logic to track whether a key already exists in the result object, simplifying handling of duplicate keys and improving code clarity. [[1]](diffhunk://#diff-31eba0e549245fa7f7c38deb81ac9bce61fd322e947fee0df818c3ef31ecf946R175-R184) [[2]](diffhunk://#diff-31eba0e549245fa7f7c38deb81ac9bce61fd322e947fee0df818c3ef31ecf946R195) [[3]](diffhunk://#diff-31eba0e549245fa7f7c38deb81ac9bce61fd322e947fee0df818c3ef31ecf946L162-R207) [[4]](diffhunk://#diff-31eba0e549245fa7f7c38deb81ac9bce61fd322e947fee0df818c3ef31ecf946L198-R242)
* Cached the result of the Latin1 charset check to avoid redundant computations during parsing. [[1]](diffhunk://#diff-31eba0e549245fa7f7c38deb81ac9bce61fd322e947fee0df818c3ef31ecf946R175-R184) [[2]](diffhunk://#diff-31eba0e549245fa7f7c38deb81ac9bce61fd322e947fee0df818c3ef31ecf946L182-R226)

### Project Configuration

* Updated `QsNet.csproj` to allow internal visibility for the new benchmarks project, enabling direct testing of internal methods.

---

### Headline
- **Internal** (`ParseQueryStringValues`): ~5.3% average speed-up across the 24 configs (median 4.9%).
  Peaks: 10–12.7% faster in the heavy cases (Count=1000, esp. `CommaLists=True`|`False` with `ValueLen=8`).
- **Public** (`Decode_Public`): ~3.6% average speed-up (median 3.2%).
  Peaks: ~10.8% faster for Count=1000 with short values.

### A few concrete deltas (baseline → current)
- Internal, 10 / False / False / 40: 2.270 μs → 2.125 μs (6.4% faster)
- Internal, 100 / True / False / 40: 25.202 μs → 23.284 μs (7.6% faster)
- Internal, 1000 / True / False / 8: 216.795 μs → 189.324 μs (12.7% faster)
- Public, 1000 / False / False / 8: 350.768 μs → 313.060 μs (10.8% faster)

### Memory
- Allocated bytes are unchanged across the board (e.g., internal 5.06 KB / 7.88 KB / … remain identical). Minor Gen counts fluctuate as usual but no net increase.

### Bottom line

Roughly ~1.06× faster internally and ~1.04× faster publicly overall, with double-digit wins on the largest workloads.

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **Refactor**
  * Faster, lower-allocation query decoding with more consistent list/parameter limit enforcement, improved comma-list and percent-encoding handling, and more reliable duplicate-parameter behavior. No public API changes.

* **Chores**
  * Added a BenchmarkDotNet project and executable to measure decode performance and memory.
  * Granted the benchmark project access to internals for accurate profiling.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->